# flavor-entertainers-backend/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Flavor Entertainers API
  description: Production-ready backend for adult entertainment performer bookings in Western Australia
  version: 1.0.0
  contact:
    name: Flavor Entertainers
    email: contact@lustandlace.com.au
  license:
    name: UNLICENSED
servers:
  - url: https://api.flavor-entertainers.example
    description: Production server
  - url: http://localhost:8080
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /payments/{booking_id}/payid-receipt:
    post:
      tags:
        - Payments
      summary: Upload PayID receipt
      description: Upload proof of PayID payment
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - receipt_file
              properties:
                receipt_file:
                  type: string
                  description: File ID of uploaded receipt
      responses:
        '200':
          description: Receipt uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
  /payments/{payment_id}/verify:
    post:
      tags:
        - Payments
      summary: Verify payment
      description: Admin verification of PayID payment receipt
      parameters:
        - name: payment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - verified
              properties:
                verified:
                  type: boolean
                notes:
                  type: string
      responses:
        '200':
          description: Payment verification completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  booking_id:
                    type: string
                  status:
                    type: string
...